pyproject.toml
[tool.poetry]
name = "todo list"
version = "0.1.0"
description = ""
authors = ["Kern"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
psycopg2 = "^2.9.10"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.28.0"
celery = "^5.4.0"
django-redis = "^5.4.0"
setuptools = "^77.0.3"
django-filter = "^25.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.2"

[tool.ruff]
line-length = 79
show-fixes = true

extend-exclude =[
    "migrations",
    "settings.py",
    "templates.py",
    "diagnostics/management/commands/_data.py",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "B",    # flake8-bugbear
    "I",    # isort
    "ICN",  # flake8-import-conventions
    "PGH",  # pygrep-hooks
    "PIE",  # flake8-pie
    "PT",   # flake8-pytest-style
    "RET",  # flake8-return
    "T20",  # flake8-print
    "UP",   # pyupgrade
    "YTT",  # flake8-2020
]
extend-ignore = ["E501"]


[tool.ruff.format]
quote-style = "double"
